requirements.txt
fastapi==0.109.0
uvicorn==0.25.0
pydantic==2.5.3
python-dotenv==1.0.0
python-multipart==0.0.6
openai-whisper==20231117
torch==2.1.2
torchaudio==2.1.2
transformers==4.36.2
google-generativeai==0.3.2
SQLAlchemy==2.0.25
passlib==1.7.4
python-jose==3.3.0
bcrypt==3.2.2
Jinja2==3.1.3
aiofiles==23.2.1
langchain==0.1.4
faiss-cpu==1.7.4
scikit-learn==1.3.2
sounddevice==0.4.6
soundfile==0.12.1
matplotlib==3.8.2
numpy==1.26.3
pandas==2.1.4
spacy==3.7.2
# WebRTC==0.0.1 (Replaced with aiortc)
alembic==1.13.1
aiortc>=1.3.2 